[ERROR HANDLER] rooms

+++ src/controllers/rooms.js
@@ -1,12 +1,11 @@
 import prisma from "../libs/prisma.js";
-
+import roomsErrorHandler from "./errors/rooms.js";
 const getRooms = async (req, res) => {
   try {
     const rooms = await prisma.room.findMany();
     return res.json(rooms);
   } catch (error) {
-    console.error("Error fetching rooms:", error);
-    return res.status(500).send("Internal Server Error");
+    roomsErrorHandler(error, res);
   }
 };
 
@@ -25,14 +24,11 @@ const getRoom = async (req, res) => {
     });
 
     if (!room) {
-      return res.status(404).send({
-        error: "Room not found",
-      });
+      throw "notFoundError";
     }
     return res.json(room);
   } catch (error) {
-    console.error("Error fetching rooms:", error);
-    return res.status(500).send("Internal Server Error");
+    roomsErrorHandler(error, res);
   }
 };
 
@@ -40,15 +36,15 @@ const postRooms = async (req, res) => {
   const { title, maxUsers, isPublic, id: hostId, filters } = req.body;
 
   try {
-    if (maxUsers < 2) throw "maxUsers must be greater than 1";
+    if (maxUsers < 2) throw "usersCountError";
     const room = await prisma.room.create({
       data: {
         title,
         maxUsers,
         isPublic,
         hostId,
-        filters:{
-          connect: filters.map(filterName => ({
+        filters: {
+          connect: filters.map((filterName) => ({
             name: filterName,
           })),
         },
@@ -61,23 +57,7 @@ const postRooms = async (req, res) => {
     });
     return res.status(201).json(room);
   } catch (error) {
-    if (error.code === "P2002") {
-      console.error("User Cann't host more than one room", hostId);
-      return res.status(409).send({
-        error: "User Cann't host more than one room",
-      });
-    } else if (error.code === "P2003") {
-      console.error("invalid User", hostId);
-      return res.status(404).send({
-        error: "invalid User",
-      });
-    } else if (error === "maxUsers must be greater than 1")
-      return res.status(409).send({
-        error: "maxUsers must be greater than 1",
-      });
-
-    console.error("Error creating room:", error);
-    return res.status(500).send({ error: "Internal Server Error" });
+    roomsErrorHandler(error, res);
   }
 };
 
@@ -85,26 +65,22 @@ const putRoom = async (req, res) => {
   const { id } = req.params;
   const { id: hostId } = req.body;
   const { title, maxUsers, isPublic, filters } = req.body;
+
   try {
+    if (maxUsers > 12 || maxUsers < 2) throw "usersCountError";
     const room = await prisma.room.findUnique({
       where: {
         id,
       },
+      include: {
+        roomUsers: true,
+      },
     });
+    if (!room) throw "notFoundError";
+    if (room.hostId !== hostId) throw "roomHostError";
+    if (maxUsers < room.roomUsers.length) throw "ExistingUsersCountError";
 
-    if (room.hostId !== hostId) {
-      return res.status(404).send({
-        error: "Room not found or you are not the host",
-      });
-    }
-
-    if (maxUsers < room.roomUsers.length) {
-      return res.status(409).send({
-        error: "maxUsers must be greater than current users",
-      });
-    }
-
-    await prisma.room.update({
+    const updatedRoom = await prisma.room.update({
       where: {
         id,
       },
@@ -115,22 +91,9 @@ const putRoom = async (req, res) => {
         filters,
       },
     });
-
-    if (!room) {
-      return res.status(404).send({
-        error: "Room not found or you are not the host",
-      });
-    }
-    return res.json(room);
+    return res.json(updatedRoom);
   } catch (error) {
-    if (error.code === "P2025") {
-      console.error("Room not found or you are not the host", id);
-      return res.status(404).send({
-        error: "Room not found or you are not the host",
-      });
-    }
-    console.log(error);
-    return res.status(500).send({ error: "Internal Server Error" });
+    roomsErrorHandler(error, res);
   }
 };
 
@@ -138,7 +101,6 @@ const deleteRoom = async (req, res) => {
   const { id } = req.params;
   const { id: hostId } = req.body;
 
-  console.log(id, hostId);
   try {
     const room = await prisma.room.findUnique({
       where: {
@@ -146,18 +108,7 @@ const deleteRoom = async (req, res) => {
       },
     });
 
-    if (!room) {
-      return res.status(404).send({
-        error: "Room not found",
-      });
-    }
-
-    if (room.hostId !== hostId) {
-      return res.status(404).send({
-        error: "you are not the host",
-      });
-    }
-
+    if (!room || room.hostId !== hostId) throw "roomHostError";
     await prisma.room.delete({
       where: {
         id,
@@ -166,8 +117,7 @@ const deleteRoom = async (req, res) => {
 
     return res.json({ message: "Room deleted successfully" });
   } catch (error) {
-    console.log(error);
-    return res.status(500).send({ error: "Internal Server Error" });
+    roomsErrorHandler(error, res);
   }
 };
 
@@ -176,11 +126,9 @@ const joinRoom = async (req, res) => {
   const { id: userId } = req.body;
 
   try {
-    if (!id || !userId) {
-      return res.status(400).send({
-        error: "id and userId are required",
-      });
-    }
+    if (!id) throw "roomIdRequiredError";
+    if (!userId) throw "userIdRequiredError";
+
     const room = await prisma.room.findUnique({
       where: {
         id,
@@ -190,17 +138,9 @@ const joinRoom = async (req, res) => {
       },
     });
 
-    if (!room) {
-      return res.status(404).send({
-        error: "Room not found",
-      });
-    }
+    if (!room) throw "notFoundError";
 
-    if (room.roomUsers.length >= room.maxUsers) {
-      return res.status(409).send({
-        error: "Room is full",
-      });
-    }
+    if (room.roomUsers.length >= room.maxUsers) throw "maxUsersError";
 
     if (room.roomUsers.find((user) => user.id === userId)) {
       return res.status(409).send({
@@ -223,8 +163,7 @@ const joinRoom = async (req, res) => {
 
     return res.json({ message: "Room joined successfully" });
   } catch (error) {
-    console.log(error);
-    return res.status(500).send({ error: "Internal Server Error" });
+    roomsErrorHandler(error, res);
   }
 };
 export default { getRooms, postRooms, getRoom, putRoom, deleteRoom, joinRoom };
